Backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.1.14
django-cors-headers==3.11.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
